fix(workflows): move the readiness hint into the canvas top-left corner [ENG-1234]
